yaml_utils: promote module to a package
The module is coupled to a project error that is really a yaml
error.  Move this into snapacraft.yaml_utils.errors and and
correspondingly for its associated unit test.
