+++ tests/unit/project/test_schema.py
@@ -23,7 +23,7 @@ from testtools.matchers import Contains, Equals
 
 # required for schema format checkers
 import snapcraft.internal.project_loader._config  # noqa: F401
-from snapcraft.project import errors
+import snapcraft.yaml_utils.errors
 from snapcraft.project._schema import Validator
 
 from . import ProjectBaseTest
@@ -58,7 +58,8 @@ class ValidationTest(ValidationBaseTest):
     def test_summary_too_long(self):
         self.data["summary"] = "a" * 80
         raised = self.assertRaises(
-            errors.YamlValidationError, Validator(self.data).validate
+            snapcraft.yaml_utils.errors.YamlValidationError,
+            Validator(self.data).validate,
         )
 
         expected_message = (
@@ -71,7 +72,8 @@ class ValidationTest(ValidationBaseTest):
         self.data["apps"] = {"service1": {}}
 
         raised = self.assertRaises(
-            errors.YamlValidationError, Validator(self.data).validate
+            snapcraft.yaml_utils.errors.YamlValidationError,
+            Validator(self.data).validate,
         )
 
         expected_message = (
@@ -86,7 +88,9 @@ class ValidationTest(ValidationBaseTest):
         mock_the_open.side_effect = FileNotFoundError()
 
         with mock.patch("snapcraft.project._schema.open", mock_the_open, create=True):
-            raised = self.assertRaises(errors.YamlValidationError, Validator, self.data)
+            raised = self.assertRaises(
+                snapcraft.yaml_utils.errors.YamlValidationError, Validator, self.data
+            )
 
         expected_message = "snapcraft validation file is missing from installation path"
         self.assertThat(raised.message, Equals(expected_message))
@@ -166,7 +170,8 @@ class ValidationTest(ValidationBaseTest):
         }
 
         raised = self.assertRaises(
-            errors.YamlValidationError, Validator(self.data).validate
+            snapcraft.yaml_utils.errors.YamlValidationError,
+            Validator(self.data).validate,
         )
 
         self.assertThat(
@@ -184,7 +189,8 @@ class ValidationTest(ValidationBaseTest):
         del self.data["adopt-info"]
 
         raised = self.assertRaises(
-            errors.YamlValidationError, Validator(self.data).validate
+            snapcraft.yaml_utils.errors.YamlValidationError,
+            Validator(self.data).validate,
         )
 
         expected_message = (
@@ -205,7 +211,7 @@ class ValidationTest(ValidationBaseTest):
 def test_daemon_dependency(data, option, value):
     data["apps"] = {"service1": {"command": "binary1", option: value}}
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     assert str(error.value).endswith(
@@ -218,7 +224,7 @@ def test_daemon_dependency(data, option, value):
 def test_required_properties(data, key):
     del data[key]
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     assert f"{key!r} is a required property" in str(error.value)
@@ -273,7 +279,7 @@ class TestInvalidNames:
     def test(self, data, name, err):
         data["name"] = name
 
-        with pytest.raises(errors.YamlValidationError) as error:
+        with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
             Validator(data).validate()
 
         assert str(error.value).endswith(
@@ -305,7 +311,7 @@ _TYPE_ENUM_TMPL = (
 def test_invalid_types(data, snap_type):
     data["type"] = snap_type
 
-    with pytest.raises(errors.YamlValidationError):
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError):
         Validator(data).validate()
 
 
@@ -319,7 +325,7 @@ def test_type_base_and_no_base(data):
 def test_type_base_and_base(data):
     data["type"] = "base"
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     assert _BASE_TYPE_MSG in str(error.value)
@@ -404,7 +410,7 @@ def test_valid_app_names(data, name):
 def test_invalid_app_names(data, name):
     data["apps"] = {name: {"command": "1"}}
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -450,7 +456,7 @@ def test_valid_refresh_modes(data, mode):
 def test_refresh_mode_daemon_missing_errors(data, mode):
     data["apps"] = {"service1": {"command": "binary1", "refresh-mode": mode}}
 
-    with pytest.raises(errors.YamlValidationError):
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError):
         Validator(data).validate()
 
 
@@ -478,7 +484,7 @@ def test_valid_modes(data, mode):
 def test_daemon_missing_errors(data, mode):
     data["apps"] = {"service1": {"command": "binary1", "stop-mode": mode}}
 
-    with pytest.raises(errors.YamlValidationError):
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError):
         Validator(data).validate()
 
 
@@ -513,7 +519,7 @@ def test_daemon_missing_errors(data, mode):
 def test_invalid_hook_names(data, name):
     data["hooks"] = {name: {"plugs": ["network"]}}
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -531,7 +537,7 @@ def test_invalid_hook_names(data, name):
 def test_invalid_part_names(data, name):
     data["parts"] = {name: {"plugin": "nil"}}
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -663,7 +669,7 @@ class TestInvalidArchitectures:
     def test(self, data, architectures, message):
         data["architectures"] = architectures
 
-        with pytest.raises(errors.YamlValidationError) as error:
+        with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
             Validator(data).validate()
 
         assert message in str(error.value)
@@ -722,7 +728,7 @@ _EXPECTED_ERROR_TEMPLATE = {
 def test_invalid_title(data, title, error_template):
     data["title"] = title
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     assert _EXPECTED_ERROR_TEMPLATE[error_template].format(title) in str(error.value)
@@ -845,7 +851,7 @@ def test_valid_version(data, version):
 def test_invalid_version(data, version):
     data["version"] = version
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -863,7 +869,7 @@ def test_invalid_version(data, version):
 def test_invalid_version_type(data):
     data["version"] = 0.1
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -880,7 +886,7 @@ def test_invalid_version_type(data):
 def test_invalid_version_length(data):
     data["version"] = "this.is.a.really.too.long.version"
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -972,7 +978,7 @@ def test_valid_compression(data, compression):
 def test_invalid_compression(data, compression):
     data["compression"] = compression
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -993,7 +999,7 @@ def test_valid_confinement(data, confinement):
 def test_invalid_confinement(data, confinement):
     data["confinement"] = confinement
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -1014,7 +1020,7 @@ def test_valid_description(data, desc):
 def test_invalid_description(data, desc):
     data["description"] = desc
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -1035,7 +1041,7 @@ def test_valid_grade(data, grade):
 def test_invalid_grade(data, grade):
     data["grade"] = grade
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -1073,7 +1079,7 @@ def test_valid_epoch(data, epoch):
 def test_invalid_epoch(data, epoch):
     data["epoch"] = epoch
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -1092,7 +1098,7 @@ def test_valid_license(data):
 def test_invalid_license(data):
     data["license"] = 1234
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = (
@@ -1113,7 +1119,7 @@ def test_valid_adapter(data, adapter):
 def test_invalid_adapter(data, adapter):
     data["apps"] = {"foo": {"command": "foo", "adapter": adapter}}
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = "The 'apps/foo/adapter' property does not match"
@@ -1127,7 +1133,7 @@ def test_invalid_adapter(data, adapter):
 def test_invalid_part_build_environment_key_type(data, build_environment):
     data["parts"]["part1"]["build-environment"] = build_environment
 
-    with pytest.raises(errors.YamlValidationError):
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError):
         Validator(data).validate()
 
 
@@ -1137,7 +1143,7 @@ def test_invalid_part_build_environment_key_type(data, build_environment):
 def test_invalid_command_chain(data, command_chain):
     data["apps"] = {"foo": {"command": "foo", "command-chain": command_chain}}
 
-    with pytest.raises(errors.YamlValidationError) as error:
+    with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
         Validator(data).validate()
 
     expected_message = "The 'apps/foo/command-chain"
@@ -1479,7 +1485,7 @@ class TestInvalidAptConfigurations:
     def test_invalid(self, data, packages, message_contains):
         data["package-repositories"] = packages
 
-        with pytest.raises(errors.YamlValidationError) as error:
+        with pytest.raises(snapcraft.yaml_utils.errors.YamlValidationError) as error:
             Validator(data).validate()
 
         assert message_contains in str(error.value)
