+++ tests/unit/project_loader/test_schema.py
@@ -23,6 +23,7 @@ import pytest
 from testscenarios.scenarios import multiply_scenarios
 from testtools.matchers import Contains, Equals
 
+import snapcraft.yaml_utils.errors
 from snapcraft import project
 from snapcraft.internal.errors import PluginError
 from snapcraft.internal.project_loader import errors, load_config
@@ -279,7 +280,9 @@ class AliasesTest(ProjectLoaderBaseTest):
     def test_invalid_alias(self):
         apps = [("test", dict(command="test", aliases=[".test"]))]
         raised = self.assertRaises(
-            project.errors.YamlValidationError, self.make_snapcraft_project, apps
+            snapcraft.yaml_utils.errors.YamlValidationError,
+            self.make_snapcraft_project,
+            apps,
         )
         expected = (
             "The {path!r} property does not match the required schema: "
